Retry makeUnique until unique username found
If there are similar usernames picked up by the `findSimilarUsernames` call, it'll throw off the count and unique username generation will get stuck on a number.

For example, a set of 4 users with usernames:
user
user1
user3

`makeUnique` would generate `user3` and fail, and never get past that point.  This fixes that by trying the count, then incrementing by 1 until it's finally unique.